% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{fetch_fixtures}
\alias{fetch_fixtures}
\title{Fetches fixture data via the fitzRoy package and filters by date range.}
\usage{
fetch_fixtures(start_date, end_date)
}
\arguments{
\item{start_date}{Minimum match date for fetched data}

\item{end_date}{Maximum match date for fetched data}
}
\description{
Fetches fixture data via the fitzRoy package and filters by date range.
}
