% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{fetch_match_results}
\alias{fetch_match_results}
\title{Fetch match results data from the Squiggle API.}
\usage{
fetch_match_results(round_number)
}
\arguments{
\item{round_number}{Fetch matches from the given round}
}
\description{
Fetch match results data from the Squiggle API.
}
