% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{fetch_player_results}
\alias{fetch_player_results}
\title{Fetches player data via the fitzRoy package and filters by date range.}
\usage{
fetch_player_results(start_date, end_date)
}
\arguments{
\item{start_date}{Minimum match date for fetched data}

\item{end_date}{Maximum match date for fetched data}
}
\description{
Fetches player data via the fitzRoy package and filters by date range.
}
