% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{match_results}
\alias{match_results}
\title{Fetches match data via the fitzRoy package and filters by date range.}
\usage{
match_results(fetch_data, start_date, end_date)
}
\arguments{
\item{fetch_data}{Whether to fetch fresh data from afltables}

\item{start_date}{Minimum match date for fetched data}

\item{end_date}{Maximum match date for fetched data}
}
\description{
Fetches match data via the fitzRoy package and filters by date range.
}
