% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betting-odds.R
\name{scrape_betting_odds}
\alias{scrape_betting_odds}
\title{Scrapes betting data for the latest round from a betting site.}
\usage{
scrape_betting_odds(splash_host)
}
\arguments{
\item{splash_host}{Hostname of the splash server to use}
}
\description{
Scrapes betting data for the latest round from a betting site.
}
