% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betting-odds.R
\name{fetch_betting_odds}
\alias{fetch_betting_odds}
\title{Fetchest betting data via the fitzRoy package and filters by date range.}
\usage{
fetch_betting_odds(start_date, end_date)
}
\arguments{
\item{start_date}{Minimum match date for fetched data}

\item{end_date}{Maximum match date for fetched data}
}
\description{
Fetchest betting data via the fitzRoy package and filters by date range.
}
