% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{fetch_match_results}
\alias{fetch_match_results}
\title{Fetches match data via the fitzRoy package and filters by date range.}
\usage{
fetch_match_results(start_date, end_date)
}
\arguments{
\item{start_date}{Minimum match date for fetched data}

\item{end_date}{Maximum match date for fetched data}
}
\description{
Fetches match data via the fitzRoy package and filters by date range.
}
