% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters.R
\name{fetch_rosters}
\alias{fetch_rosters}
\title{Scrapes team roster data (i.e. which players are playing for each team) for
a given round from afl.com.au, cleans it, and returns it as a dataframe.}
\usage{
fetch_rosters(browser)
}
\arguments{
\item{browser}{Selenium browser object for navigating to pages and crawling the DOM.}
}
\description{
Scrapes team roster data (i.e. which players are playing for each team) for
a given round from afl.com.au, cleans it, and returns it as a dataframe.
}
