% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters.R
\name{fetch_rosters}
\alias{fetch_rosters}
\title{Scrapes team roster data (i.e. which players are playing for each team) for
a given round from afl.com.au, cleans it, and returns it as a dataframe.}
\usage{
fetch_rosters(round_number)
}
\arguments{
\item{round_number}{Fetch the rosters from this round. Required,
because it is assigned to the round_number column of the returned data set.}
}
\description{
Scrapes team roster data (i.e. which players are playing for each team) for
a given round from afl.com.au, cleans it, and returns it as a dataframe.
}
